ModelBaseEcon Reference

Options

ModelBaseEcon.OptionsMod — Module

OptionsMod

Sub-module of ModelBaseEcon, although it can be used independently. Implements the Options data structure.

Contents

Options

getoption - read the value of an option

getoption! - if not present, also create an option

setoption! - create or update the value of an option

ModelBaseEcon.OptionsMod.Options — Type

Options

A collection of key-value pairs representing the options controlling the behaviour or the definition of a Model object. The key is the option name and is always a Symbol, or converted to Symbol, while the value can be anything.

The options can be accessed using dot notation. Functions getoption and setoption! are also provided. They can be used for programmatic processing of options as well as when the option name is not a valid Julia identifier.

See also: Options, getoption, getoption!, setoption!

Examples

julia> o = Options(maxiter=20, tol=1e-7)
Options:
    maxiter=20
    tol=1.0e-7

julia> o.maxiter = 25
25

julia> o
Options:
    maxiter=25
    tol=1.0e-7

ModelBaseEcon.OptionsMod.Options — Method

Options(::Options)

Construct an Options instance as an exact copy of an existing instance.

ModelBaseEcon.OptionsMod.Options — Method

Options(key=value, ...)
Options(:key=>value, ...)

Construct an Options instance with key-value pairs given as keyword arguments or as a list of pairs. If the latter is used, each key must be a Symbol.

ModelBaseEcon.OptionsMod.getoption — Function

getoption(o::Options; name=default [, name=default, ...])
getoption(o::Options, name, default)

Retrieve the value of an option or a set of options. The provided defaults are used when the option doesn't exit.

The return value is the value of the option requested or, if the option doesn't exist, the default. In the first version of the function, if there are more than one options requested, the return value is a tuple.

In the second version, the name could be a symbol or a string, which can be helpful if the name of the option is not a valid identifier.

ModelBaseEcon.OptionsMod.getoption! — Function

getoption!(o::Options; name=default [, name=default, ...])
getoption!(o::Options, name, default)

Retrieve the value of an option or a set of options. If the name does not match an existing option, the Options instance is updated by inserting the given name and default value.

The return value is the value of the option requested (or the default). In the first version of the function, if there are more than one options requested, the return value is a tuple.

In the second version, the name could be a symbol or a string, which can be helpful if the name of the option is not a valid identifier.

ModelBaseEcon.OptionsMod.setoption! — Function

setoption!(o::Options; name=default [, name=default, ...])
setoption!(o::Options, name, default)

Retrieve the value of an option or a set of options. If the name does not match an existing option, the Options instance is updated by inserting the given name and default value.

The return value is the value of the option requested (or the default). In the first version of the function, if there are more than one options requested, the return value is a tuple.

In the second version, the name could be a symbol or a string, which can be helpful if the name of the option is not a valid identifier.

Timer

ModelBaseEcon.Timer — Module

Timer

A sub-module of ModelBaseEcon, although it can be used independently. Provides functionality for measuring the aggregate time spent in individual lines or blocks of code.

Tip

The implementation here is quick-and-dirty and is intended for getting a rough idea of where the bottlenecks are. It is best used for timing blocks of code that are passed through relatively few times and each pass takes a relatively long time. In the opposite extreme case (fast code that is called many times), the current implementation of @timer might add extreme overhead.

Contents

inittimer - Enable collection of timer data.

stoptimer - Disable collection of timer data.

printtimer - Display timer data.

@timer - Measure the runtime taken by the given code.

Example

julia> true
[...]

ModelBaseEcon.Timer.inittimer — Method

inittimer()

Enable the collection of timing data. Existing timing data is lost. By default, collection of timing data is disabled.

See also: stoptimer, @timer, printtimer

ModelBaseEcon.Timer.printtimer — Function

printtimer(io::IO=Base.stdout)

Display timing data.

Timing data is displayed in a table with each row containing the number of calls, total time in seconds, and the source line or block tag. Rows are sorted in order of decreasing total time.

See also: @timer

ModelBaseEcon.Timer.stoptimer — Method

stoptimer()

Disable the collection of timing data. Existing data is lost. By default, collection of timing data is disabled.

See also: inittimer, @timer, printtimer

ModelBaseEcon.Timer.@timer — Macro

@timer(code)
@timer(tag::String, code)

Measure the number of calls and the total time taken by the given code.

If a tag string is not provided, one is generated from the source file and line. The return value of this macro call is the return value of the code.

Warning

Important limitation is that the code must not contain a return, break, continue, or any other jump out of it. If it does, the program would run correctly, but the timing data collected would be incorrect.

See also: inittimer, stoptimer, printtimer

Example

julia> inittimer()

julia> @timer Base.sleep(1.0)

julia> printtimer()
[...]

ModelBaseEcon

ModelBaseEcon.ModelBaseEcon — Module

ModelBaseEcon

This package is part of the StateSpaceEcon ecosystem. It contains the basic elements needed for model definition. StateSpaceEcon works with model objects defined with ModelBaseEcon.

ModelBaseEcon.Equation — Type

struct Equation <: AbstractEquation

Data structure representing a single equation in our state space model.

Methods (for users)

TODO

Implementation (for developers)

During the phase of definition of the Model, this type simply stores the expression entered by the user. During @initialize(), the true data structure is constructed. We need this, because the construction of the equation requaires information from the Model object, which may not be available at the time the equation expression is first read.

ModelBaseEcon.LinearizationError — Type

LinearizationError <: ModelErrorBase

A concrete error type used when a model cannot be linearized for some reason.

ModelBaseEcon.Model — Type

Model <: AbstractModel

Data structure that represents a macroeconomic state space model.

ModelBaseEcon.ModelError — Type

struct ModelError <: ModelErrorBase

Concrete error type used when no specific error description is available.

ModelBaseEcon.ModelErrorBase — Type

ModelErrorBase

Abstract error type, base for specific error types used in ModelBaseEcon.

Implementation (note for developers)

When implementing a derived error type, override two functions:

msg(e::SomeModelError) returning a string with the error message;

hint(e::SomeModelError) returning a string containing a suggestion of how to fix the problem. Optional, if not implemented for a type, the fallback implementation returns an empty string.

ModelBaseEcon.ModelNotInitError — Type

struct ModelNotInitError <: ModelErrorBase

Specific error type used when there's an attempt to use a Model object that has not been initialized.

ModelBaseEcon.NotImplementedError — Type

struct NotImplementedError <: ModelErrorBase

Specific error type used when a feature is planned but not yet implemented.

ModelBaseEcon.SteadyStateData — Type

SteadyStateData

Data structure that holds information about the steady state solution of the Model.

ModelBaseEcon.SteadyStateEquation — Type

struct SteadyStateEquation <: AbstractEquation

Data structure representing an individual steady state equation.

ModelBaseEcon.alleqns — Method

alleqns(ssd::SteadyStateData)

Return a list of all steady state equations.

The list contains all equations derived from the dynamic system and all explicitly added steady state constraints.

ModelBaseEcon.eval_R! — Function

eval_R!(res::AbstractArray{Float64,1}, point::AbstractArray{Float64, 2}, ::MED) where MED <: AbstractModelEvaluationData

Evaluate the model residual at the given point using the given model evaluation structure. The residual is stored in the provided vector.

Implementation details (for developers)

When creating a new type of model evaluation data, you must define a method of this function specialized to it.

The point argument will be a 2d array, with the number of rows equal to maxlag+maxlead+1 and the number of columns equal to the number of variables+shocks+auxvars of the model. The res vector will have the same length as the number of equations + auxiliary equations. Your implementation must not modify point and must update res.

See also: eval_RJ

ModelBaseEcon.eval_RJ — Function

eval_RJ(point::AbstractArray{Float64, 2}, ::MED) where MED <: AbstractModelEvaluationData

Evaluate the model residual and its Jacobian at the given point using the given model evaluation structure. Return a tuple, with the first element being the residual and the second element being the Jacobian.

Implementation details (for developers)

When creating a new type of model evaluation data, you must define a method of this function specialized to it.

The point argument will be a 2d array, with the number of rows equal to maxlag+maxlead+1 and the number of columns equal to the number of variables+shocks+auxvars of the model. Your implementation must not modify point and must return the tuple of (residual, Jacobian) evaluated at the given point. The Jacobian is expected to be SparseMatrixCSC (this might change in the future).

See also: eval_R!

ModelBaseEcon.geteqn — Method

geteqn(i, ssd::SteadyStateData)

Return the i-th steady state equation. Index i is interpreted as in the output of alleqns. Calling geteqn(i, sdd) has the same effect as alleqn(ssd)[i], but it's more efficient.

Example

# Iterate all equations like this:
for i = 1:neqns(ssd)
    eqn = geteqn(i, ssd)
    # do something awesome with `eqn` and `i`
end

ModelBaseEcon.islinearized — Method

islinearized(m::Model)

Return true if the given model is linearized and false otherwise.

ModelBaseEcon.issssolved — Method

issssolved(sstate::SteadyStateData)

Return true if the steady state has been solved, or false otherwise.

ModelBaseEcon.linearize! — Function

linearize!(model::Model; <keyword arguments>)

Transform model into its linear approximation about its steady state.

Keyword arguments

sstate - linearize about the provided steady state solution

deviation::Bool - whether or not the linearized model will tread data passed

to is as deviation from the steady state

See also: linearized and with_linearized

ModelBaseEcon.linearized — Method

linearized(model::Model; <arguments>)

Create a new model that is the linear approximation of the given model about its steady state.

Keyword arguments

sstate - linearize about the provided steady state solution

deviation::Bool - whether or not the linearized model will tread data passed

to is as deviation from the steady state

See also: linearize! and with_linearized

ModelBaseEcon.neqns — Method

neqns(ssd::SteadyStateData)

Return the total number of equations in the steady state system, including the ones derived from the dynamic system and the ones added explicitly as steady state constraints.

ModelBaseEcon.printsstate — Method

printsstate([io::IO,] ssd::SteadyStateData)

Display steady state solution.

Steady state solution is presented in a table, where the first column is the name of the variable, the second and third columns are the corresponding values of the level and the slope. If the value is not determined (as per its mask value) then it is displayed as "*".

ModelBaseEcon.with_linearized — Method

with_linearized(F::Function, model::Model; <arguments>)

Apply the given function on a new model that is the linear approximation of the given model about its steady state. This is meant to be used with the do syntax, as in the example below.

Keyword arguments

sstate - linearize about the provided steady state solution

deviation::Bool - whether or not the linearized model will tread data passed

to is as deviation from the steady state

See also: linearize! and with_linearized

Example

with_linearized(m) do lm
    # do something awesome with linearized model `lm`
end
# model `m` is still non-linear.

ModelBaseEcon.@autoexogenize — Macro

@autoexogenize model begin
    varname = shkname
    ...
end

Define a mapping between variables and shocks that can be used to conveniently swap exogenous and endogenous variables.

ModelBaseEcon.@autoshocks — Macro

@autoshocks model

Create a list of shocks that matches the list of variables. Each shock name is created from a variable name by appending "_shk".

ModelBaseEcon.@equations — Macro

Usage example:

@equations model begin
    y[t] = a * y[t-1] + b * y[t+1] + y_shk[t]

ModelBaseEcon.@initialize — Macro

@initialize model

Prepare a model instance for analysis. Call this macro after all variable names, shock names and equations have been defined.

ModelBaseEcon.@load_example — Macro

Deprecated. Use @using_example instead.

ModelBaseEcon.@parameters — Macro

@parameters model begin
    name = value
    ...
end

Declare and define the model parameters.

The parameters must have values. Provide the information in a series of assignment statements wrapped inside a begin-end block. The names can be used in equations as if they were regular variables.

ModelBaseEcon.@shocks — Macro

@shocks model names...
@shocks model begin
    names...
end

Define the names of transition shocks in the model.

Example

```jldoctest @shocks model ashk bshk c_shk

If the list is long, use a begin-end block separating names with newline or semicolon

@shocks model begin ashk; bshk c_shk end ````

ModelBaseEcon.@steadystate — Macro

@steadystate model [type] equation

Add a steady state equation to the model.

The steady state system of the model is automatically derived from the dynamic system. Use this macro to define additional equations for the steady state. This is particularly useful in the case of a non-linear model that might have multiple steady state, or the steady state might be difficult to solve for, to help the steady state solver find the one you want to use.

model is the model instance you want to update

type (optional) is the type of constraint you want to add. This can be level

or slope. If missing, the default is level

equation is the expression defining the steady state constraint. In the

equation, use variables and shocks from the model, but without any t-references.

ModelBaseEcon.@using_example — Macro

@using_example name

Load models from the package examples/ folder. The @load_example version is deprecated - stop using it now.

ModelBaseEcon.@variables — Macro

@variables model names...
@variables model begin
    names...
end

Define the names of transition variables in the model.

Example

```jldoctest @variables model a b c

If the list is long, use a begin-end block separating names with newline or semicolon

@variables model begin a; b c end ````