TimeSeriesEcon Reference

TimeSeriesEcon.TimeSeriesEcon — Module

TimeSeriesEcon

This package is part of the StateSpaceEcon ecosystem. TimeSeriesEcon.jl provides functionality to work with low-Frequency discrete macroeconomic time-series data.

Frequencies (abstract type):

Unit

Monthly

Quarterly

Yearly

Types:

MIT{Frequency} (aka "Moment In Time")

a primitive type denoting monthly, quarterly, and yearly dates

TSeries{Frequency}

an AbstractVector that can be indexed using MIT

Functions:

MIT Constructors/Functions

mm(year::Int, period::Int): returns a monthly MIT type instance

qq(year::Int, period::Int): returns a quarterly MIT type instance

yy(year::Int): returns a yearly MIT type instance

ii(x::Int): returns a unit MIT type instance

year(x::MIT): returns a Int64 year value associated with x

period(x::MIT): returns a Int64 period value associated with x

frequencyof(x::MIT): returns <: Frequency assosicated wtih x

Functions operating on TSeries

mitrange(x::TSeries): returns a UnitRange{MIT{Frequency}} for the given x

firstdate(x::TSeries): returns MIT{Frequency} first date associated with x

lastdate(x::TSeries): returns MIT{Frequency} last date associated with x

ppy(x::TSeries): returns the number of periods per year for x::TSeries. (ppy also accepts x::MIT and x::Frequency)

shift(x::TSeries, i::Int64): shifts the dates of x by firstdate(x) - i

shift!: in-place version of shift

pct(x::TSeries, shift_value::Int64; islog::Bool = false): calculates percent rate of change of x::TSeries

apct(x::TSeries, islog::Bool = false): calculates annualized percent rate of change of x::TSeries

nanrm!(x::TSeries, type::Symbol=:both): removes NaN from x::TSeries

TimeSeriesEcon.Frequency — Type

Frequency

Frequency is an abstract type.

Monthly, Quarterly, Yearly, Unit abstract types are subtypes of Frequency.

Examples

julia> Quarterly <: Frequency
true

TimeSeriesEcon.MIT — Type

MIT{Frequency}(x::Int64)

MIT is a primitive type (Signed 64) representing a discrete date.

Note: Please use yy, qq, 'mm', ii to instantiate MITs. MITs are mainly used internally for the package development.

Examples

julia> MIT{Yearly}(2000)
2000Y
julia> MIT{Monthly}(2020*12 + 3)
2020M4

TimeSeriesEcon.Monthly — Type

Monthly

See also: Frequency ```

TimeSeriesEcon.Quarterly — Type

Quarterly

See also: Frequency ```

TimeSeriesEcon.TSeries — Type

struct TSeries{Frequency} <: AbstractVector{Float64}

Data structure representing a time-series vector. The following operations are allowed:

indexing using MIT (aka "moment-in-time") and UnitRange{MIT}

assignment using MIT and UnitRange{MIT}

In addition, most of the operations available to Julia vectors (+, -, *, etc.) are supported by TSeries as well.

Examples

Create TSeries

julia> x = TSeries(qq(2020, 1), ones(4))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

Index into TSeries

julia> x[qq(2020, 1)]
1.0

julia> x[qq(2020, 1):qq(2020, 2)]
TSeries{Quarterly} of length 2
2020Q1: 1.0
2020Q2: 1.0

Assignment using MIT

julia> x[qq(2020, 1)] = 100; x
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

julia> x[qq(2020, 1):qq(2020, 2)] = 100; x
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 100.0
2020Q3: 1.0
2020Q4: 1.0

Arithmetic Operations on TSeries

julia> x = TSeries(qq(2020, 1), ones(4))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

julia> 2*x + 98
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 100.0
2020Q3: 100.0
2020Q4: 100.0

julia> log(exp(x))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

TimeSeriesEcon.Unit — Type

Unit

See also: Frequency ```

TimeSeriesEcon.Yearly — Type

Yearly

See also: Frequency ```

Base.:* — Method

3 * ts
ts * 3 returns timeseries with every element multiplied by 3

Base.:/ — Method

ts / 3 returns timeseries with every element divided by 3

Base.diff — Method

diff(x::TSeries, k::Int64 = -1)

Same as Iris implementation of diff

Base.getindex — Method

getindex using Vector{MIT}. Note the difference between Vector{MIT} and UnitRange{MIT}

Base.getindex — Method

getindex using UnitRange{MIT}

Base.getindex — Method

getindex using MIT

Base.hcat — Method

Horizonatal Concatenation of `TSeries`

Examples

julia> a = TSeries(ii(1), ones(3));
julia> b = TSeries(ii(2), ones(3));
julia> [a b]
4×2 Array{Float64,2}:
   1.0  NaN
   1.0    1.0
   1.0    1.0
 NaN      1.0

Base.setindex! — Method

setindex a vector using UnitRange{MIT}

Base.setindex! — Method

setindex a value using UnitRange{MIT}

Base.setindex! — Method

setindex a value using MIT

Base.setindex! — Method

setindex values from other TSeries using UnitRange{MIT}

Base.setindex! — Method

setindex values from other TSeries using MIT

TimeSeriesEcon.apct — Function

apct(x::TSeries, islog::Bool)

Calculate annualised percent rate of change in x.

Note: The implementation is similar to IRIS.

Examples

julia> x = TSeries(qq(2018, 1), Vector(1:8));

julia> apct(x)
TSeries{Quarterly} of length 7
2018Q2: 1500.0
2018Q3: 406.25
2018Q4: 216.04938271604937
2019Q1: 144.140625
2019Q2: 107.35999999999999
2019Q3: 85.26234567901243
2019Q4: 70.59558517284461

See also: pct

TimeSeriesEcon.firstdate — Method

firstdate(x::TSeries)

Return an MIT indicating the first date in the TSeries.

Examples

julia> firstdate(TSeries(qq(2020, 1), ones(10)))
2020Q1

TimeSeriesEcon.frequencyof — Method

frequencyof(::MIT)
frequencyof(::Type{MIT})

Return the Frequency type of the given MIT instance of type.

TimeSeriesEcon.frequencyof — Method

Can be applied to a TSeries instance or a range of MIT to return its Frequency.

TimeSeriesEcon.ii — Method

ii(x::Int64)

Represents an Integer date and return MIT{Unit} type instance

Examples

julia> ii(123)
ii(123)

julia> ii(123) + 5
ii(128)

julia> typeof(ii(123))
MIT{Unit}

TimeSeriesEcon.lastdate — Method

lastdate(x::TSeries)

Return an MIT indicating the last date in the TSeries.

Examples

julia> lastdate(TSeries(qq(2020, 1), ones(10)))
2022Q2

TimeSeriesEcon.leftcropnan! — Method

leftcropnan!(x::TSeries)

Remove NaN values from starting at the beginning of x, in-place.

Note: an internal function.

TimeSeriesEcon.mitrange — Method

mitrange(x::TSeries)

Return an UnitRange{MIT{<:Frequency}} associated with x.

Examples

julia> mitrange(TSeries(qq(2020, 1), ones(4)))
2020Q1:2020Q4

TimeSeriesEcon.mm — Method

mm(y::Int64, p::Int64)

Represents a Monthly date and returns MIT{Monthly} type instance.

Examples

julia> mm(2020, 1)
2020M1

julia> mm(2020, 1) + 5
2020M6

TimeSeriesEcon.nanrm! — Function

nanrm!(s::TSeries, type::Symbol)

Remove NaN values that are either at the beginning of the s and/or end of x.

Examples

julia> s = TSeries(yy(2018), [NaN, NaN, 1, 2, NaN]);

julia> nanrm!(s);

julia> s
TSeries{Yearly} of length 2
2020Y: 1.0
2021Y: 2.0

TimeSeriesEcon.pct — Method

pct(x::TSeries, shift_value::Int64, islog::Bool)

Calculate percentage growth in x given a shift_value.

Note: The implementation is similar to IRIS.

Examples

julia> x = TSeries(yy(2000), Vector(1:4));

julia> pct(x, -1)
TSeries{Yearly} of length 3
2001Y: 100.0
2002Y: 50.0
2003Y: 33.33333333333333

See also: apct

TimeSeriesEcon.period — Method

period(x::MIT)

Return ::Int64 representing period for a given MIT

Examples

julia> period(mm(2020, 1))
1

TimeSeriesEcon.ppy — Method

ppy(::MIT)
ppy(::Type{MIT})

When applied to an MIT instance or type, return the ppy of its frequency.

TimeSeriesEcon.ppy — Method

ppy(::TSeries)
ppy(::Type{TSeries})

When applied to a TSeries instance or type, return the ppy of its frequency.

TimeSeriesEcon.ppy — Method

ppy(::Frequency)

Returns the number of periods per year for a given Frequency, MIT, and TSeries

Examples

julia> ppy(Quarterly)                   # Frequency
4

julia> ppy(mm(2020, 1))                 # MIT
12

julia> ppy(TSeries(yy(2020), ones(3)))   # TSeries
1

TimeSeriesEcon.qq — Method

qq(y::Int64, p::Int64)

Represents a Quarterly date and returns MIT{Quarterly} type instance.

Examples

julia> qq(2020, 1)
2020Q1

julia> qq(2020, 1) + 5
2021Q2

TimeSeriesEcon.rightcropnan! — Method

rightcropnan!(x::TSeries)

Remove NaN values from the end of x

Note: an internal function.

TimeSeriesEcon.shift! — Method

shift!(x::TSeries, n::Int64)

Shift dates of x back by k periods, in-place. Note: The implementation of is similar to IRIS ts{1}.

Examples

julia> x = TSeries(qq(2020, 1), ones(4));

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0

See also: shift

TimeSeriesEcon.shift — Method

shift(x::TSeries, n::Int64)

Shift dates of x back by k periods. Note: The implementation of is similar to IRIS ts{1}.

Examples

julia> shift(TSeries(qq(2020, 1), ones(4)), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0

julia> shift(TSeries(qq(2020, 1), ones(4)), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0
2021Q1: 1.0

See also: shift!

TimeSeriesEcon.year — Method

year(x::MIT)

Return ::Int64 representing year for a given MIT

Note: an internal method for now.

Examples

julia> year(mm(2020, 1))
2020

TimeSeriesEcon.yy — Function

yy(y::Int64)

Represents a Yearly date and returns MIT{Yearly} type instance.

Examples

julia> yy(2020)
2020Y

julia> yy(2020) + 5
2025Y

TimeSeriesEcon.@rec — Macro

@rec(eqn, rng)

Computes recursive calculations for the given eqn and rng.

Examples

julia> s = TSeries(ii(1), zeros(1));

julia> # Initial values

julia> s[ii(1)] = 0;

julia> s[ii(2)] = 1;

julia> @rec s[t] = s[t-1] + s[t-2] ii(3):ii(10)

julia> s
TSeries{Unit} of length 10
ii(1): 0.0
ii(2): 1.0
ii(3): 1.0
ii(4): 2.0
ii(5): 3.0
ii(6): 5.0
ii(7): 8.0
ii(8): 13.0
ii(9): 21.0
ii(10): 34.0