TimeSeriesEcon Reference

TimeSeriesEcon.Duration — Type

MIT{F <: Frequency}, Duration{F <: Frequency}

Two types representing a moment in time (like 2020Q1 or 2020-01-01) and duration (the quantity of time between two moments).

Both of these have a Frequency as a type parameter and both internally are represented by integer values.

If you imagine a time axis of the given Frequency, MIT values are ordinal (correspond to points) while Duration values are cardinal (correspond to distances).

TimeSeriesEcon.Frequency — Type

abstract type Frequency end

The abstract supertype for all frequencies. See also: Unit and YPFrequency

TimeSeriesEcon.MIT — Type

MIT{F <: Frequency}, Duration{F <: Frequency}

Two types representing a moment in time (like 2020Q1 or 2020-01-01) and duration (the quantity of time between two moments).

Both of these have a Frequency as a type parameter and both internally are represented by integer values.

If you imagine a time axis of the given Frequency, MIT values are ordinal (correspond to points) while Duration values are cardinal (correspond to distances).

TimeSeriesEcon.MIT — Method

MIT{F}(year, period) where {F <: YPFrequency}

Construct an MIT instance from year and period. This is valid only for frequencies subtyped from YPFrequency.

TimeSeriesEcon.Monthly — Type

Yearly, Quarterly, Monthly

Frequencies corresponding to 1, 4, and 12 periods per year. See also: YPFrequency

TimeSeriesEcon.Quarterly — Type

Yearly, Quarterly, Monthly

Frequencies corresponding to 1, 4, and 12 periods per year. See also: YPFrequency

TimeSeriesEcon.TSeries — Type

mutable struct TSeries{F, T, C} <: AbstractVector{T}
    firstdate::MIT{F}
    values::C
end

Time series with frequency F with values of type T stored in a container of type C. By default the type is Float64 and the container is Vector{Float64}.

Construction: ts = TSeries(args...)

The standard construction is `TSeries(firstdate::MIT, values::AbstractVector)`

If the first argument is an MIT-range (instead or an MIT), then the length
of the `values` container must match the length of the given range.

In the case of a range argument, the `values` can be omitted, in which case
the container is initializes with `undef`. Or you can also pass a constant
and then the `values` will be filled with that constant. To accomplish this,
you can also use `fill`, e.g., `TSeries(20Q1:20Q4, 5)` is the same as
`fill(5, 20Q1:20Q4)`.

If only a `firstdate::MIT` is given, the `values` container is initialized
to an empty `Vector`.

If only an `n::Integer` is given, it is the same as passing the range
`0U .+ (1:n)`. An initialization argument is not allowed in this case.

A `TSeries` can also be constructed with `copy`, `similar`, and `fill`.

Indexing:

Indexing with an `MIT` or a range of `MIT` works as you'd expect.

Indexing with `Integer`s works the same as with `Vector`.

Indexing with `Bool`-array works as you'd expect. For example,
`s[s .< 0.0] .*= -1` multiplies in place the negative entries of `s` by -1,
so effectively it's the same as `s .= abs.(s)`.

There are important differences between indexing with MIT and not
using MIT (i.e., using Integer or Bool-array).

* with MIT-range we return a TSeries with the given range, otherwise we
  return a `Vector`

* the range can be extended (the TSeries resized appropriately) by assigning
  outside the current range. This works only with MIT (you get a BoundsError
  if you try to assign outside the Integer range).

* `begin` and `end` are MIT, so either use both or none of them. For example
  `s[2:end]` doesn't work because 2 is an `Int` and `end` is an `MIT`. You
  should use `s[begin+1:end]`.

TimeSeriesEcon.Unit — Type

struct Unit <: Frequency end

Represents a non-dimensional frequency (not associated with the calendar). See also: YPFrequency

TimeSeriesEcon.Workspace — Type

struct Workspace … end

A collection of variables.

TimeSeriesEcon.YPFrequency — Type

abstract type YPFrequency{N} <: Frequency end;

Represents a calendar frequency based on a number of periods in a year. The type parameter N is the number of periods and must be a positive integer. See also: Yearly, Quarterly, Monthly

TimeSeriesEcon.Yearly — Type

Yearly, Quarterly, Monthly

Frequencies corresponding to 1, 4, and 12 periods per year. See also: YPFrequency

TimeSeriesEcon.apct — Method

apct(x::TSeries, islog::Bool)

Calculate annualised percent rate of change in x.

Note: The implementation is similar to IRIS.

Examples

julia> x = TSeries(qq(2018, 1), Vector(1:8));

julia> apct(x)
TSeries{Quarterly} of length 7
2018Q2: 1500.0
2018Q3: 406.25
2018Q4: 216.04938271604937
2019Q1: 144.140625
2019Q2: 107.35999999999999
2019Q3: 85.26234567901243
2019Q4: 70.59558517284461

See also: pct

TimeSeriesEcon.compare — Function

@compare x y [options] 
compare(x, y [; options])

Compare two Workspace recursively and print out the differences. MVTSeries and Dict with keys of type Symbol are treated like Workspace. TSeries and other Vector are compared using isapprox, so feel free to supply rtol or atol.

Optional argument name can be used for the top name. Default is "_".

Parameter showequal=true causes the report to include objects that are the same. Default behaviour, with showequal=false, is to report only the differences.

Parameter ignoremissing=true causes objects that appear in one but not the other workspace to be ignored. That is, they are not printed and do not affect the return value true or false. Default is ignoremissing=false meaning they will be printed and return value will be false.

TimeSeriesEcon.fconvert — Method

fconvert(F, t)

Convert the time series t to the desired frequency F.

TimeSeriesEcon.fconvert — Method

fconvert(F1, t::TSeries{F2}; method) where {F1 <: YPFrequency, F2 <: YPFrequency}

Convert between frequencies of the YPFrequency variety.

TODO: describe method when converting to a higher frequency (interpolation) TODO: describe method when converting to a lower frequency (aggregation)

TimeSeriesEcon.firstdate — Function

firstdate(ts), lastdate(ts)

Return the first and last date of the allocated data for the given TSeries. These are identical to firstindex and lastindex.

TimeSeriesEcon.frequencyof — Function

frequencyof(x), frequencyof(T)

Return the Frequency type of the given value x or type T.

TimeSeriesEcon.lag — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.lag! — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.lastdate — Function

firstdate(ts), lastdate(ts)

Return the first and last date of the allocated data for the given TSeries. These are identical to firstindex and lastindex.

TimeSeriesEcon.lead — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.lead! — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.mit2yp — Function

mit2yp(x::MIT)

Recover the year and period from a given MIT value. This is valid only if the frequency is subtyped from YPFrequency.

TimeSeriesEcon.mm — Function

mm(year, period), qq(year, period), yy(year, period=1)

IRIS type constructors for MIT values with frequencies that have year and period.

TimeSeriesEcon.moving — Function

moving(x, n)

Compute the moving average of x over a window of n periods. If n > 0 the window is backward-looking (-n+1:0) and if n < 0 the window is forward-looking (0:-n-1).

TimeSeriesEcon.overlay — Method

overlay(rng, t1, t2, ...)

If the first argument is a range (must be of the same frequency), that becomes the range of the resulting TSeries.

TimeSeriesEcon.overlay — Method

overlay(t1, t2, ...)

Construct a TSeries in which each observation is taken from the first non-missing observation in the list of arguments. A missing observation is one for which istypenan returns true.

All TSeries in the argument list must be of the same frequency. The data type of the resulting TSeries is computed by the standard promotion of numerical types in Julia. Its range is the union of the ranges of the arguments.

TimeSeriesEcon.period — Function

year(x::MIT), period(x::MIT)

Return the year and period of an MIT value x. This only makes sense if the frequency of x is subtyped from YPFrequency.

TimeSeriesEcon.qq — Function

mm(year, period), qq(year, period), yy(year, period=1)

IRIS type constructors for MIT values with frequencies that have year and period.

TimeSeriesEcon.rangeof — Method

rangeof(s)

Return the stored range of the given time series object.

TimeSeriesEcon.rangeof — Method

rangeof(s; drop::Integer)

Return the stored range of s adjusted by dropping drop periods. If drop is positive, we drop from the beginning and if drop is negative we drop from the end. This adds convenience when using @rec

Example

julia> q = TSeries(20Q1:21Q4); rangeof(q; drop=1)
20Q2:21Q4

julia> rangeof(q; drop=-4)
20Q1:20Q4

julia> q[begin:begin+1] .= 1; @rec rangeof(q; drop=2) q[t] = q[t-1] + q[t-2]; q
8-element TSeries{Quarterly} with range 20Q1:21Q4:
    20Q1 : 1.0
    20Q2 : 1.0
    20Q3 : 2.0
    20Q4 : 3.0
    21Q1 : 5.0
    21Q2 : 8.0
    21Q3 : 13.0
    21Q4 : 21.0

TimeSeriesEcon.shift — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.shift! — Function

shift(x, n)
shift!(x, n)
lag(x, n=1)
lag!(x, n=1)
lead(x, n=1)
lead!(x, n=1)

Shift, lag or lead the TSeries x by n periods. By convention shift is the same as lead while lag(x,n) is the same as shift(x, -n). The versions ending in ! do it in place, while the others create a new TSeries instance.

Examples

julia> shift(TSeries(2020Q1, 1:4), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

julia> shift(TSeries(2020Q1, 1:4), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 2.0
2020Q4: 3.0
2021Q1: 4.0

julia> x = TSeries(2020Q1, 1:4);

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 2.0
2020Q2: 3.0
2020Q3: 4.0

TimeSeriesEcon.strip! — Method

strip!(w::Workspace; recursive=true)

Apply strip! to all TSeries members of the given workspace. This includes nested workspaces, unless recursive=false.

TimeSeriesEcon.typenan — Function

typenan(x), typenan(T)

Return a value that indicates Not-A-Number of the same type as the given x or of the given type T.

For floating point types, this is the IEEE-defined NaN. For integer types, we use typemax(). This is not ideal, but it'll do for now.

TimeSeriesEcon.undiff — Function

undiff(dvar, [date => value])
undiff!(var, dvar; fromdate=firstdate(dvar)-1)

Inverse of diff, i.e. var remains unchanged under undiff!(var, diff(var)) or undiff(diff(var), firstdate(var)=>first(var)). This is the same as cumsum, but specific to time series.

In the case of undiff the second argument is an "anchor" Pair specifying a known value at some time period. Typically this will be the period just before the first date of dvar, but doesn't have to be. If the date falls outside the rangeof(dvar) we extend dvar with zeros as necessary. If missing, this argument defaults to firstdate(dvar)-1 => 0.

In the case of undiff!, the var argument provides the "anchor" value and the storage location for the result. The fromdate parameter specifies the date of the "anchor" and the anchor value is taken from var. See important note below.

The in-place version (undiff!) works only with TSeries. The other version (undiff) works with MVTSeries as well as TSeries. In the case of MVTSeries the anchor value must be a Vector, or a Martix with 1 row, of the same length as the number of columns of dvar.

Note

In the case of undiff! the meaning of parameter fromdate is different from the meaning of date in the second argument of undiff. This only matters if fromdate falls somewhere in the middle of the range of dvar.

In the case of undiff!, all values of dvar at, and prior to, fromdate are ignored (considered zero). Effectively, values of var up to, and including, fromdate remain unchanged.

By contrast, in undiff with date => value somewhere in the middle of the range of dvar, the operation is applied over the full range of dvar, both before and after date, and then the result is adjusted by adding or subtracting a constant such that in the end we have result[date]=value.

TimeSeriesEcon.undiff! — Function

undiff(dvar, [date => value])
undiff!(var, dvar; fromdate=firstdate(dvar)-1)

Inverse of diff, i.e. var remains unchanged under undiff!(var, diff(var)) or undiff(diff(var), firstdate(var)=>first(var)). This is the same as cumsum, but specific to time series.

In the case of undiff the second argument is an "anchor" Pair specifying a known value at some time period. Typically this will be the period just before the first date of dvar, but doesn't have to be. If the date falls outside the rangeof(dvar) we extend dvar with zeros as necessary. If missing, this argument defaults to firstdate(dvar)-1 => 0.

In the case of undiff!, the var argument provides the "anchor" value and the storage location for the result. The fromdate parameter specifies the date of the "anchor" and the anchor value is taken from var. See important note below.

The in-place version (undiff!) works only with TSeries. The other version (undiff) works with MVTSeries as well as TSeries. In the case of MVTSeries the anchor value must be a Vector, or a Martix with 1 row, of the same length as the number of columns of dvar.

Note

In the case of undiff! the meaning of parameter fromdate is different from the meaning of date in the second argument of undiff. This only matters if fromdate falls somewhere in the middle of the range of dvar.

In the case of undiff!, all values of dvar at, and prior to, fromdate are ignored (considered zero). Effectively, values of var up to, and including, fromdate remain unchanged.

By contrast, in undiff with date => value somewhere in the middle of the range of dvar, the operation is applied over the full range of dvar, both before and after date, and then the result is adjusted by adding or subtracting a constant such that in the end we have result[date]=value.

TimeSeriesEcon.year — Function

year(x::MIT), period(x::MIT)

Return the year and period of an MIT value x. This only makes sense if the frequency of x is subtyped from YPFrequency.

TimeSeriesEcon.ytypct — Method

ytypct(x)

Year-to-year percent change in x.

TimeSeriesEcon.yy — Function

mm(year, period), qq(year, period), yy(year, period=1)

IRIS type constructors for MIT values with frequencies that have year and period.

TimeSeriesEcon.@compare — Macro

@compare x y [options] 
compare(x, y [; options])

Compare two Workspace recursively and print out the differences. MVTSeries and Dict with keys of type Symbol are treated like Workspace. TSeries and other Vector are compared using isapprox, so feel free to supply rtol or atol.

Optional argument name can be used for the top name. Default is "_".

Parameter showequal=true causes the report to include objects that are the same. Default behaviour, with showequal=false, is to report only the differences.

Parameter ignoremissing=true causes objects that appear in one but not the other workspace to be ignored. That is, they are not printed and do not affect the return value true or false. Default is ignoremissing=false meaning they will be printed and return value will be false.

TimeSeriesEcon.@rec — Macro

@rec [index=]range expression

Computes recursive operations on time series. The first argument is the range and the second argument is an expression to be evaluated over that range.

The expression is meant to be an assignment, but it doesn't have to be.

The the range can specify an optional indexing variable (as in a for loop). If not given, the variable is assumed to be t.

Examples

julia> s = TSeries(1U)
Empty TSeries{Unit} starting 5U

julia> s[1U] = s[2U] = 1; s
2-element TSeries{Unit} with range 1U:2U:
      1U : 1.0
      2U : 1.0

julia> @rec t=3U:10U s[t] = s[t-1] + s[t-2]

julia> s
10-element TSeries{Unit} with range 1U:10U:
      1U : 1.0
      2U : 1.0
      3U : 2.0
      4U : 3.0
      5U : 5.0
      6U : 8.0
      7U : 13.0
      8U : 21.0
      9U : 34.0
     10U : 55.0